enhance flow and flow vault connection management commands
