+++ docs/auth0_flows_vault.md
@@ -9,6 +9,6 @@ Manage the vault connections that store credentials for flow integrations.
 
 ## Commands
 
-- [auth0 flows vault connections](auth0_flows_vault_connections.md) - Manage Flow vault connections
+- [auth0 flows vault connections](auth0_flows_vault_connections.md) - Manage Flow vault connections.
 - [auth0 flows vault open](auth0_flows_vault_open.md) - Open the Vault in the Auth0 Dashboard
