+++ docs/auth0_flows_vault_open.md
@@ -36,7 +36,7 @@ auth0 flows vault open [flags]
 
 ## Related Commands
 
-- [auth0 flows vault connections](auth0_flows_vault_connections.md) - Manage Flow vault connections
+- [auth0 flows vault connections](auth0_flows_vault_connections.md) - Manage Flow vault connections.
 - [auth0 flows vault open](auth0_flows_vault_open.md) - Open the Vault in the Auth0 Dashboard
