+++ internal/auth/auth.go
@@ -133,7 +133,9 @@ var RequiredScopes = []string{
 	"read:phone_templates",
 	"create:email_templates", "read:email_templates", "update:email_templates",
 	"create:email_provider", "read:email_provider", "update:email_provider", "delete:email_provider",
-	"read:flows", "create:flows", "read:forms", "create:forms", "update:forms", "delete:forms", "read:flows_vault_connections",
+	"read:flows", "create:flows", "update:flows", "delete:flows",
+	"read:forms", "create:forms", "update:forms", "delete:forms",
+	"read:flows_vault_connections", "create:flows_vault_connections", "update:flows_vault_connections", "delete:flows_vault_connections",
 	"read:connections", "update:connections", "read:connections_options", "update:connections_options",
 	"read:client_keys", "read:logs", "read:tenant_settings", "update:tenant_settings",
 	"read:custom_domains", "create:custom_domains", "update:custom_domains", "delete:custom_domains",
