+++ test/integration/flows-vault-test-cases.yaml
@@ -1,0 +1,100 @@
+config:
+  inherit-env: true
+  retries: 1
+
+tests:
+  001 - it successfully lists all vault connections (json):
+    command: auth0 flows vault connections list --json
+    exit-code: 0
+
+  002 - it prints the setup template for a vault connection:
+    command: auth0 flows vault connections create --setup-template --app-id HTTP
+    exit-code: 0
+    stdout:
+      contains:
+        - '"setup"'
+        - '"type"'
+
+  003 - it prints the vault app builder URL for open:
+    command: auth0 flows vault open HTTP --no-input
+    exit-code: 0
+    stderr:
+      contains:
+        - forms.auth0.com
+        - /vault/apps/HTTP/edit
+
+  004 - it fails to create a vault connection without required flags:
+    command: auth0 flows vault connections create --no-input
+    exit-code: 1
+    stderr:
+      contains:
+        - app id is required
+
+  005 - it successfully creates a vault connection from a fixture file (json):
+    command: auth0 flows vault connections create --setup-file ./test/integration/fixtures/vault-connection.json --name integration-test-connection-created --app-id HTTP --no-input --json
+    exit-code: 0
+    stdout:
+      json:
+        name: "integration-test-connection-created"
+
+  006 - it successfully lists all vault connections with data:
+    command: auth0 flows vault connections list
+    exit-code: 0
+    stdout:
+      contains:
+        - ID
+        - NAME
+        - APP
+
+  007 - given a test vault connection, it successfully shows the connection details:
+    command: auth0 flows vault connections show $(./test/integration/scripts/get-vault-connection-id.sh)
+    exit-code: 0
+    stdout:
+      contains:
+        - ID
+        - NAME
+        - integration-test-connection
+
+  008 - given a test vault connection, it does not echo the setup secret on show:
+    command: auth0 flows vault connections show $(./test/integration/scripts/get-vault-connection-id.sh)
+    exit-code: 0
+    stdout:
+      not-contains:
+        - integration-test-token
+
+  009 - given a test vault connection, it successfully shows the connection details (json):
+    command: auth0 flows vault connections show $(./test/integration/scripts/get-vault-connection-id.sh) --json
+    exit-code: 0
+    stdout:
+      json:
+        name: "integration-test-connection"
+
+  010 - given a test vault connection, it successfully updates the connection name:
+    command: auth0 flows vault connections update $(./test/integration/scripts/get-vault-connection-id.sh) --name integration-test-connection-updated --json
+    exit-code: 0
+    stdout:
+      json:
+        name: "integration-test-connection-updated"
+
+  011 - given a test vault connection, it successfully updates the connection from a fixture file:
+    command: auth0 flows vault connections update $(./test/integration/scripts/get-vault-connection-id.sh) -f ./test/integration/fixtures/update-vault-connection.json --json
+    exit-code: 0
+    stdout:
+      json:
+        name: "integration-test-connection-fixture-updated"
+
+  012 - agent mode refuses to delete a vault connection without force:
+    command: AUTH0_AGENT_MODE=true auth0 flows vault connections delete $(./test/integration/scripts/get-vault-connection-id.sh)
+    exit-code: 1
+    stderr:
+      contains:
+        - destructive command
+        - --force
+
+  013 - given a test vault connection, it successfully deletes the connection:
+    command: auth0 flows vault connections delete $(./test/integration/scripts/get-vault-connection-id.sh) --force
+    exit-code: 0
+
+  014 - it cleans up all vault connections created by this suite:
+    command: ./test/integration/scripts/cleanup-vault-connections.sh
+    exit-code: 0
